--- a/includes/classes/Command.php
+++ b/includes/classes/Command.php
@@ -905,7 +905,7 @@ class Command extends WP_CLI_Command {
 
 			$loop_counter++;
 			if ( ( $loop_counter % 10 ) === 0 ) {
-				$time_elapsed_diff = isset( $time_elapsed ) ? ' (+' . (string) ( timer_stop( 0, 2 ) - $time_elapsed ). ')' : '';
+				$time_elapsed_diff = isset( $time_elapsed ) ? ' (+' . (string) ( timer_stop( 0, 2 ) - $time_elapsed ) . ')' : '';
 				$time_elapsed      = timer_stop( 0, 2 );
 				WP_CLI::log( WP_CLI::colorize( '%Y' . esc_html__( 'Time elapsed: ', 'elasticpress' ) . '%N' . $time_elapsed . $time_elapsed_diff ) );