--- a/includes/classes/Command.php
+++ b/includes/classes/Command.php
@@ -900,7 +900,8 @@ class Command extends WP_CLI_Command {
 				break;
 			}
 
-			$last_processed_object_id = $query['objects'][ array_key_last( $query['objects'] ) ]->ID;
+			$last_object_array_key    = array_keys( $query['objects'] )[ count( $query['objects'] ) - 1 ];
+			$last_processed_object_id = $query['objects'][ $last_object_array_key ]->ID;
 			WP_CLI::log( sprintf( esc_html__( 'Processed %1$d/%2$d. Last Object ID: %3$d', 'elasticpress' ), (int) ( $synced + count( $failed_objects ) ), (int) $total_objects_to_process, (int) $last_processed_object_id ) );
 
 			$loop_counter++;