# Purpose

Pagination with SQL's offset can get considerably slow on large sites as it has to tread through all the rows until it finds where to pick up. This PR switches away from using offset and instead ranges of post IDs which allows it to execute queries much faster. I've also prevented the need for running `SQL_CALC_FOUND_ROWS` on each query. There are comparisons of the query changes and performance profiling at the end.

To accomplish this new way of pagination while still using WP_Query, I'm passing custom query args and then filtering on `posts_where` based on these args.

# Testing

At the moment, I've kept the functionality behind a feature flag to make testing/comparing easier. To test, you'll want to apply this PR but then also add the `[--advanced-pagination] [--start-post-id]` CLI args to the docblock in `search/includes/classes/commands/class-corecommand.php` to allow them to passthrough the `wp vip-search index` CLI if you want to use that.

Ideally this just happens by default when it's possible, and we remove the feature flag & related logic when it's merged.

```
wp vip-search index --version=2 --indexables=post --advanced-pagination --start-post-id=2500000
```

I'm not 100% sure on the "end range" functionality you'll see in the code, and would appreciate some advice there. Mainly, is it worth it? Removing the end range appears to actually (very slightly) speed up the query and makes the code a bit simpler, but having the end-range does highly limit the rows scanned. Could go even further on limiting rows scanned, as right now there is an arbitrary limit of 50,000 IDs. In testing, changing this to lower doesn't really have an effect - it's already very fast.

# Profiling

```
### CURRENT STATUS

explain SELECT SQL_CALC_FOUND_ROWS  wp_posts.* FROM wp_posts  WHERE 1=1  AND wp_posts.post_type IN (XXX) AND ((wp_posts.post_status = 'publish' OR wp_posts.post_status = 'future' OR wp_posts.post_status = 'trash' OR wp_posts.post_status = 'inherit' OR wp_posts.post_status = 'request-pending' OR wp_posts.post_status = 'request-confirmed' OR wp_posts.post_status = 'request-failed' OR wp_posts.post_status = 'request-completed' OR wp_posts.post_status = 'spam' OR wp_posts.post_status = 'pitch' OR wp_posts.post_status = 'draft' OR wp_posts.post_status = 'needs-editing' OR wp_posts.post_status = 'ready-to-publish' OR wp_posts.post_status = 'sync-only' OR wp_posts.post_status = 'private'))  ORDER BY wp_posts.ID DESC LIMIT 1500000, 500;

+------+-------------+----------+-------+------------------+---------+---------+------+---------+-------------+
| id   | select_type | table    | type  | possible_keys    | key     | key_len | ref  | rows    | Extra       |
+------+-------------+----------+-------+------------------+---------+---------+------+---------+-------------+
|    1 | SIMPLE      | wp_posts | index | type_status_date | PRIMARY | 8       | NULL | 5714850 | Using where |
+------+-------------+----------+-------+------------------+---------+---------+------+---------+-------------+

In the following tests, I had to drop to 250 posts per page instead of 500 (timeouts otherwise) and I removed SQL_CALC_FOUND_ROWS for a more fair comparison due to how I changed the logic in this PR.
Otherwise, you could expect 60+ second queries that often timeout. You'll see how offset gets progressively worse even with these adjustments:

Offset 2500: 0.35 secs
Offset 25000: 1.54 secs
Offset 250000: 23.42 secs
Offset 2500000: 35.66 sec

### WITH THIS PR

explain SELECT wp_posts.* FROM wp_posts  WHERE 1=1 AND wp_posts.ID < 2598793 AND wp_posts.ID > 2548793  AND wp_posts.post_type IN (XXX) AND ((wp_posts.post_status = 'publish' OR wp_posts.post_status = 'future' OR wp_posts.post_status = 'trash' OR wp_posts.post_status = 'inherit' OR wp_posts.post_status = 'request-pending' OR wp_posts.post_status = 'request-confirmed' OR wp_posts.post_status = 'request-failed' OR wp_posts.post_status = 'request-completed' OR wp_posts.post_status = 'spam' OR wp_posts.post_status = 'pitch' OR wp_posts.post_status = 'draft' OR wp_posts.post_status = 'needs-editing' OR wp_posts.post_status = 'ready-to-publish' OR wp_posts.post_status = 'sync-only' OR wp_posts.post_status = 'private'))  ORDER BY wp_posts.ID DESC LIMIT 0, 500;

+------+-------------+----------+-------+--------------------------+---------+---------+------+-------+-------------+
| id   | select_type | table    | type  | possible_keys            | key     | key_len | ref  | rows  | Extra       |
+------+-------------+----------+-------+--------------------------+---------+---------+------+-------+-------------+
|    1 | SIMPLE      | wp_posts | range | PRIMARY,type_status_date | PRIMARY | 8       | NULL | 81838 | Using where |
+------+-------------+----------+-------+--------------------------+---------+---------+------+-------+-------------+

 Actual Query Time: 0.28 sec. No noticeable fluctuation with differing IDs.

Perhaps worth noting that query time is 0.27 sec if I remove the end range (wp_posts.ID > 2548793), though then 2857433 rows are scanned.
```