@WPprodigy 2020-10-13
Addressed the feedback. Some notable changes:

- Cleaned up and made this behavior the default, along with proper support for passing in `--start-object-id` along with `--end-object-id`. (Will now actually index the start id rather than the one after it)
- Reworked the query arg logic so that `last_processed_object_id` is now what we pass between loops. This will allow for greater flexibility in the future, such as if we wanted to switch up the order to ASC for example. All of the related logic for the ranges now lives in `bulk_indexing_filter_posts_where()` and there alone (quite abstract, could be used for terms later on).
- Logging enhancements.

I abandoned the artificial end ranges for now, due to complexities with how the index_helper() loop is structured / how it decided when to stop looping. Still not completely sure why queries through $wpdb don't trigger the MAX_JOIN_SIZE error, but seems to be alright and query time is actually the slightest bit faster without the end range for whatever reason.